Furniture components (with workshop now as component)
