+++ Assets/Scripts/Models/Buildable/Components/Component.cs.meta
@@ -1,7 +1,7 @@
 fileFormatVersion: 2
-guid: 0ddae74203c0b714886f7baaab95276b
-timeCreated: 1473636178
-licenseType: Pro
+guid: f7551857c47bb6c448a8ad53caae9705
+timeCreated: 1474919672
+licenseType: Free
 MonoImporter:
   serializedVersion: 2
   defaultReferences: []
