+++ Assets/Scripts/Models/Buildable/Components/Workshop.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: 4b21914c6f5f4274b9268f99607053d7
+timeCreated: 1474919672
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
