+++ Assets/Scripts/Models/Buildable/Furniture.cs
@@ -46,11 +46,8 @@ public class Furniture : IXmlSerializable, ISelectable, IPrototypable, IContextA
     /// These context menu lua action are used to build the context menu of the furniture.
     /// </summary>
     private List<ContextMenuLuaAction> contextMenuLuaActions;
-
-    /// <summary>
-    /// Workshop reference if furniture is consumer/producer (not null). 
-    /// </summary>
-    private FurnitureWorkshop workshop;
+    
+    private HashSet<ProjectPorcupine.Buildable.Components.Component> components;
 
     // This is the generic type of object this is, allowing things to interact with it based on it's generic type
     private HashSet<string> typeTags;
@@ -93,6 +90,7 @@ public class Furniture : IXmlSerializable, ISelectable, IPrototypable, IContextA
         Width = 1;
         DragType = "single";
         LinksToNeighbour = string.Empty;
+        components = new HashSet<ProjectPorcupine.Buildable.Components.Component>();
     }
 
     /// <summary>
@@ -117,7 +115,8 @@ public class Furniture : IXmlSerializable, ISelectable, IPrototypable, IContextA
 
         Parameters = new Parameter(other.Parameters);
         Jobs = new BuildableJobs(this, other.Jobs);
-        workshop = other.workshop; // don't need to clone here, as all are prototype things (not changing)
+        // don't need to clone here, as all are prototype things (not changing)
+        components = new HashSet<ProjectPorcupine.Buildable.Components.Component>(other.components);
 
         if (other.Animation != null)
         {
@@ -189,16 +188,7 @@ public class Furniture : IXmlSerializable, ISelectable, IPrototypable, IContextA
     /// </summary>
     /// <value>The Color of the furniture.</value>
     public Color Tint { get; set; }
-
-    /// <summary>
-    /// Returns whether this instance is workshop or not.
-    /// </summary>
-    /// <value><c>true</c> if this instance is workshop; otherwise, <c>false</c>.</value>
-    public bool IsWorkshop
-    {
-        get { return workshop != null; }
-    }
-
+        
     /// <summary>
     /// Gets or sets a value indicating whether the door is Vertical or not.
     /// Should be false if the furniture is not a door.
@@ -383,13 +373,14 @@ public class Furniture : IXmlSerializable, ISelectable, IPrototypable, IContextA
 
         // We know our placement destination is valid.
         Furniture obj = proto.Clone();
-        obj.Tile = tile;        
-        if (obj.IsWorkshop)
-        {
-            // need to update reference to furniture for workshop (is there a nicer way?)
-            obj.workshop.SetParentFurniture(obj);
-        }
+        obj.Tile = tile;
 
+        // need to update reference to furniture
+        foreach (var comp in obj.components)
+        {
+            comp.ParentFurniture = obj;
+        }
+      
         if (tile.PlaceFurniture(obj) == false)
         {
             // For some reason, we weren't able to place our object in this tile.
@@ -480,11 +471,14 @@ public class Furniture : IXmlSerializable, ISelectable, IPrototypable, IContextA
             EventActions.Trigger("OnUpdate", this, deltaTime);
         }
 
-        if (IsWorkshop && IsBeingDestroyed == false)
+        if (!IsBeingDestroyed)
         {
-            workshop.Update(deltaTime);
+            foreach (var cmp in components)
+            {
+                cmp.Update(deltaTime);
+            }
         }
-
+        
         if (Animation != null)
         {
             Animation.Update(deltaTime);
@@ -710,10 +704,14 @@ public class Furniture : IXmlSerializable, ISelectable, IPrototypable, IContextA
                     reader.Read();
                     UnlocalizedDescription = reader.ReadContentAsString();
                     break;
-                case "Workshop":                   
-                    workshop = FurnitureWorkshop.Deserialize(reader);
-                    workshop.SetParentFurniture(this);
-                    workshop.Initialize();
+                case "Component":
+                    var cmp = ProjectPorcupine.Buildable.Components.Component.Deserialize(reader);
+                    if (cmp != null)
+                    {
+                        cmp.ParentFurniture = this;
+                        cmp.Initialize();
+                        components.Add(cmp);
+                    }
                     break;
             }
         }
@@ -991,11 +989,16 @@ public class Furniture : IXmlSerializable, ISelectable, IPrototypable, IContextA
 
     public IEnumerable<string> GetAdditionalInfo()
     {
-        if (IsWorkshop)
+        // try to get some info from components
+        foreach(var comp in components)
         {
-            yield return workshop.GetDescription();
+            string desc = comp.GetDescription();
+            if(!string.IsNullOrEmpty(desc))
+            {
+                yield return desc;
+            }
         }
-
+        
         yield return string.Format("Hitpoint 18 / 18");
 
         if (PowerConnection != null)
@@ -1086,15 +1089,19 @@ public class Furniture : IXmlSerializable, ISelectable, IPrototypable, IContextA
             }
         }
 
-        // context menu if it's workshop and has multiple production chains
-        if (IsWorkshop && workshop.WorkshopMenuActions != null)
+        // check for context menus of components
+        foreach(var comp in components)
         {
-            foreach (WorkshopContextMenu factoryContextMenuAction in workshop.WorkshopMenuActions)
+            var compContextMenu = comp.GetContextMenu();
+            if (compContextMenu != null)
             {
-                yield return CreateWorkshopContextMenuItem(factoryContextMenuAction);
+                foreach (ComponentContextMenu compContextMenuAction in compContextMenu)
+                {
+                    yield return CreateWorkshopContextMenuItem(compContextMenuAction);
+                }
             }
-        }       
-
+        }
+       
         foreach (ContextMenuLuaAction contextMenuLuaAction in contextMenuLuaActions)
         {
             if (!contextMenuLuaAction.DevModeOnly ||
@@ -1183,13 +1190,13 @@ public class Furniture : IXmlSerializable, ISelectable, IPrototypable, IContextA
         return true;
     }
 
-    private ContextMenuAction CreateWorkshopContextMenuItem(WorkshopContextMenu factoryContextMenuAction)
+    private ContextMenuAction CreateWorkshopContextMenuItem(ComponentContextMenu componentContextMenuAction)
     {
         return new ContextMenuAction
         {
-            Text = factoryContextMenuAction.ProductionChainName, // TODO: localization here
+            Text = componentContextMenuAction.Name, // TODO: localization here
             RequireCharacterSelected = false,
-            Action = (cma, c) => InvokeContextMenuAction(factoryContextMenuAction.Function, factoryContextMenuAction.ProductionChainName)
+            Action = (cma, c) => InvokeContextMenuAction(componentContextMenuAction.Function, componentContextMenuAction.Name)
         };
     }
