+++ Assets/StreamingAssets/Data/Furniture.xml
@@ -392,7 +392,7 @@
 
         <JobWorkSpotOffset X="1" Y="0" />
 
-        <Workshop>
+        <Component type="Workshop">
             <ProductionChain name="Copper smelting" processingTime="3">
                 <Input>
                     <Item objectType="Raw Copper" amount="2" slotPosX="0" slotPosY="0" />
@@ -410,7 +410,7 @@
                 </Output>
             </ProductionChain>
             <UsedAnimations idle="idle" running="running" />
-        </Workshop>
+        </Component>
 
         <Animations>
             <Animation state="idle" fps="2" looping="false">
@@ -477,7 +477,7 @@
             <Inventory type="Uranium" amount="1" />
         </DeconstructJob>
 
-      <Workshop>
+      <Component type="Workshop">
           <ProductionChain name="Power Cell Pressing" processingTime="5">
               <Input>
                   <Item objectType="Uranium" amount="3" slotPosX="0" slotPosY="0" />
@@ -487,7 +487,7 @@
                   <Item objectType="Power Cell" amount="1" slotPosX="2" slotPosY="1" />
               </Output>
           </ProductionChain>
-      </Workshop>
+      </Component>
 
         <LocalizationCode>furn_power_cell_press</LocalizationCode>
         <UnlocalizedDescription>furn_power_cell_press_desc</UnlocalizedDescription>
